CS5330_F24_Group4_Term_Project-main/RealTimeTranslationApp/requirements.txt
Flask
numpy
opencv-python
easyocr
googletrans
Pillow
certifi
tailwindcss